Support base patch in bundled pyinstaller executable
